+++ changelog.md
@@ -1,6 +1,8 @@
 ## 04 Jul 24
 - Fix antispam not respecting the restric to servers constant
 
+- Make /access-blueprint response try to send some parts of message via text and some via files when the complete text is too big
+
 ## 03 Jul 24
 - Make bot send part of the /access-blueprint command response as a regular message if a blueprint is posted in a blueprints channel
   - Move part of /access-blueprint command code in its own function